requirements.txt
pyrogram==2.0.59
aiohttp
tgcrypto
Pillow
lottie